# gateway/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "gateway microservice of youtube-clone app.",
        "title": "rest api of gateway service",
        "contact": {
            "name": "rzaf",
            "url": "https://www.github.com/rzaf/youtube-clone"
        }
    },
    "basePath": "/api",
    "paths": {
        "/comments/medias/{url}": {
            "get": {
                "description": "get comments of media",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "comments"
                ],
                "summary": "get comments of media",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "items perpage",
                        "name": "perpage",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "newest",
                            "oldest",
                            "most-liked",
                            "least-liked",
                            "most-disliked",
                            "least-disliked",
                            "most-replied",
                            "least-replied"
                        ],
                        "type": "string",
                        "default": "newest",
                        "description": "sort type",
                        "name": "sort",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "url",
                        "name": "url",
                        "in": "path"
                    },
                    {
                        "type": "string",
                        "description": "optional authentication",
                        "name": "X-API-KEY",
                        "in": "header"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "204": {
                        "description": "no content",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "request failed",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "create comment/reply",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "comments"
                ],
                "summary": "create comment/reply",
                "parameters": [
                    {
                        "type": "string",
                        "description": "description",
                        "name": "description",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "reply_url",
                        "name": "reply_url",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "media url",
                        "name": "url",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "request failed",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/comments/{commentUrl}": {
            "get": {
                "description": "get comment",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "comments"
                ],
                "summary": "get comment",
                "parameters": [
                    {
                        "type": "string",
                        "description": "commentUrl",
                        "name": "commentUrl",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "optional authentication",
                        "name": "X-API-KEY",
                        "in": "header"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "request failed",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "edit comment",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "comments"
                ],
                "summary": "edit comment",
                "parameters": [
                    {
                        "type": "string",
                        "description": "description",
                        "name": "description",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "commentUrl",
                        "name": "commentUrl",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "request failed",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "not authenticated",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "403": {
                        "description": "not authorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "delete comment",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "comments"
                ],
                "summary": "delete comment",
                "parameters": [
                    {
                        "type": "string",
                        "description": "commentUrl",
                        "name": "commentUrl",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "request failed",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "not authenticated",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "403": {
                        "description": "not authorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/comments/{commentUrl}/replies": {
            "get": {
                "description": "get replies of comments",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "comments"
                ],
                "summary": "get replies of comments",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "items perpage",
                        "name": "perpage",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "newest",
                            "oldest",
                            "most-liked",
                            "least-liked",
                            "most-disliked",
                            "least-disliked",
                            "most-replied",
                            "least-replied"
                        ],
                        "type": "string",
                        "default": "newest",
                        "description": "sort type",
                        "name": "sort",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "commentUrl",
                        "name": "commentUrl",
                        "in": "path"
                    },
                    {
                        "type": "string",
                        "description": "optional authentication",
                        "name": "X-API-KEY",
                        "in": "header"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "204": {
                        "description": "no content",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "request failed",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "api key of user",
            "type": "apiKey",
            "name": "X-API-KEY",
            "in": "header"
        }
    }
}